Update of DashboardService, DomainMatricsProjection and OperatorMatricsProjection

+++ src/main/java/org/generation/italy/model/repositories/OperatorMatricsProjection.java
@@ -1,7 +1,7 @@
 package org.generation.italy.model.repositories;
 
 public interface OperatorMatricsProjection {
-    Long getOperatorId();
+    Integer getOperatorId();
     String getOperatorName();
     Long getTotalMinutes();
     Long getRegistrationCount();
